building CRUD updated

+++ backend/src/main/java/cs309_dorm_backend/service/BuildingServiceImpl.java
@@ -5,25 +5,39 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Optional;
 
 import cs309_dorm_backend.dao.BuildingRepo;
 
 @Service
 public class BuildingServiceImpl implements BuildingService {
+
     @Autowired
     private BuildingRepo buildingRepo;
 
     @Override
-    public void deleteById(long id) {
-        buildingRepo.deleteById(id);
+    public List<Building> findAll() {
+        return buildingRepo.findAll();
     }
 
     @Override
-    public Building save(Building Building) {
-        return null;
+    public Building findById(int id) {
+        return buildingRepo.findById(id).orElse(null);
     }
 
-    public List<Building> findAll() {
-        return buildingRepo.findAll();
+    @Override
+    public boolean deleteById(int id) {
+        Optional<Building> building = buildingRepo.findById(id);
+        if (building.isPresent()) {
+            buildingRepo.deleteById(id);
+            return true;
+        } else {
+            return false;
+        }
+    }
+
+    @Override
+    public Building save(Building building) {
+        return buildingRepo.save(building);
     }
 }
